feat: Implement editing of PeriodicChoreEntry inside of calendar (#91)
* feat: Implement editing of PeriodicChoreEntry inside of calendar

By clicking on an entry in the calendar, an EditEntryModal is opened
which allows editing of that entry. So far, the date is displayed read-
only, while the member who completed the chore is fully changeable.
Clicking "Save" updates the PeriodicChore.

* fix(components): Fix inactive member not displayed in EditEntryModal

When an entry is edited whose member was set to inactive in the
meantime, the member selection showed as empty (the invalid state),
even though the member obviously still exists. By introducing a special
case during filtering, this is circumvented.