feat: Notify client of new version availability and show modal
The hash of the running script is compared to the hash of the script
currently available on disk (reported by the server via WebSocket), and
if there is a mismatch, a modal is shown to the user prompting them to
reload the page. This is meant to avoid incompatibilities between
outdated clients interacting with up-to-date servers.