fix(frontend): Use TailwindCSS for all remaining components
This patch eliminates all but two CSS files in favor of Tailwind
classes. The remaining files are `NavigationBarLayout.css` and
`NavigationBar.css`, which are rather hard to fully convert at this
point due to the weird way in which they share a CSS variable (namely,
`--nav-thickness`) as well as media queries.